Partially organized, still need to fix errors
