+++ src/main/java/frc/robot/Robot.java
@@ -29,167 +29,156 @@ import org.littletonrobotics.junction.wpilog.WPILOGReader;
 import org.littletonrobotics.junction.wpilog.WPILOGWriter;
 
 /**
- * The VM is configured to automatically run this class, and to call the
- * functions corresponding to
- * each mode, as described in the TimedRobot documentation. If you change the
- * name of this class or
- * the package after creating this project, you must also update the
- * build.gradle file in the
+ * The VM is configured to automatically run this class, and to call the functions corresponding to
+ * each mode, as described in the TimedRobot documentation. If you change the name of this class or
+ * the package after creating this project, you must also update the build.gradle file in the
  * project.
  */
 public class Robot extends LoggedRobot {
-    private RobotContainer robotContainer;
+  private RobotContainer robotContainer;
 
-    private Command autoCommand;
-    private boolean matchStartingMethodCalled = false;
+  private Command autoCommand;
+  private boolean matchStartingMethodCalled = false;
 
-    public Robot() {
-        Pathfinding.setPathfinder(new LocalADStarAK());
+  public Robot() {
+    Pathfinding.setPathfinder(new LocalADStarAK());
 
-        PathPlannerLogging.setLogTargetPoseCallback(
-                (pose) -> Logger.recordOutput("PathPlanner/TargetPose", pose));
-        PathPlannerLogging.setLogCurrentPoseCallback(
-                (pose) -> Logger.recordOutput("PathPlanner/CurrentPose", pose));
-        PathPlannerLogging.setLogActivePathCallback(
-                (path) -> Logger.recordOutput("PathPlanner/ActivePath", path.toArray(new Pose2d[path.size()])));
+    PathPlannerLogging.setLogTargetPoseCallback(
+        (pose) -> Logger.recordOutput("PathPlanner/TargetPose", pose));
+    PathPlannerLogging.setLogCurrentPoseCallback(
+        (pose) -> Logger.recordOutput("PathPlanner/CurrentPose", pose));
+    PathPlannerLogging.setLogActivePathCallback(
+        (path) ->
+            Logger.recordOutput("PathPlanner/ActivePath", path.toArray(new Pose2d[path.size()])));
 
-        // Record metadata
-        Logger.recordMetadata("ProjectName", BuildConstants.MAVEN_NAME);
-        Logger.recordMetadata("BuildDate", BuildConstants.BUILD_DATE);
-        Logger.recordMetadata("GitSHA", BuildConstants.GIT_SHA);
-        Logger.recordMetadata("GitDate", BuildConstants.GIT_DATE);
-        Logger.recordMetadata("GitBranch", BuildConstants.GIT_BRANCH);
-        switch (BuildConstants.DIRTY) {
-            case 0:
-                Logger.recordMetadata("GitDirty", "All changes committed");
-                break;
-            case 1:
-                Logger.recordMetadata("GitDirty", "Uncomitted changes");
-                break;
-            default:
-                Logger.recordMetadata("GitDirty", "Unknown");
-                break;
+    // Record metadata
+    Logger.recordMetadata("ProjectName", BuildConstants.MAVEN_NAME);
+    Logger.recordMetadata("BuildDate", BuildConstants.BUILD_DATE);
+    Logger.recordMetadata("GitSHA", BuildConstants.GIT_SHA);
+    Logger.recordMetadata("GitDate", BuildConstants.GIT_DATE);
+    Logger.recordMetadata("GitBranch", BuildConstants.GIT_BRANCH);
+    switch (BuildConstants.DIRTY) {
+      case 0:
+        Logger.recordMetadata("GitDirty", "All changes committed");
+        break;
+      case 1:
+        Logger.recordMetadata("GitDirty", "Uncomitted changes");
+        break;
+      default:
+        Logger.recordMetadata("GitDirty", "Unknown");
+        break;
+    }
+
+    // Set up data receivers & replay source
+    switch (Constants.getRobotMode()) {
+      case REAL:
+        // Running on a real robot, log to a USB stick ("/U/logs")
+        Logger.addDataReceiver(new WPILOGWriter());
+        Logger.addDataReceiver(new NT4Publisher());
+        break;
+
+      case SIM:
+        // Running a physics simulator, log to NT
+        Logger.addDataReceiver(new NT4Publisher());
+        break;
+
+      case REPLAY:
+        // Replaying a log, set up replay source
+        setUseTiming(false); // Run as fast as possible
+        String logPath = LogFileUtil.findReplayLog();
+
+        if (logPath == null || logPath.isEmpty()) {
+          System.err.println("Error: Replay log not found. Ensure a valid log file is available.");
+          throw new IllegalStateException("Replay log not found.");
         }
-
-        // Set up data receivers & replay source
-        switch (Constants.getRobotMode()) {
-            case REAL:
-                // Running on a real robot, log to a USB stick ("/U/logs")
-                Logger.addDataReceiver(new WPILOGWriter());
-                Logger.addDataReceiver(new NT4Publisher());
-                break;
-
-            case SIM:
-                // Running a physics simulator, log to NT
-                Logger.addDataReceiver(new NT4Publisher());
-                break;
-
-            case REPLAY:
-                // Replaying a log, set up replay source
-                setUseTiming(false); // Run as fast as possible
-                String logPath = LogFileUtil.findReplayLog();
-
-                if (logPath == null || logPath.isEmpty()) {
-                    System.err.println("Error: Replay log not found. Ensure a valid log file is available.");
-                    throw new IllegalStateException("Replay log not found.");
-                }
-                Logger.setReplaySource(new WPILOGReader(logPath));
-                Logger.addDataReceiver(new WPILOGWriter(LogFileUtil.addPathSuffix(logPath, "_sim")));
-                break;
-        }
-
-        // Start AdvantageKit logger
-        Logger.start();
-
-        robotContainer = new RobotContainer();
-
-        FollowPathCommand.warmupCommand().schedule();
-        PathfindingCommand.warmupCommand().schedule();
+        Logger.setReplaySource(new WPILOGReader(logPath));
+        Logger.addDataReceiver(new WPILOGWriter(LogFileUtil.addPathSuffix(logPath, "_sim")));
+        break;
     }
 
-    /** This function is called periodically during all modes. */
-    @Override
-    public void robotPeriodic() {
-        Threads.setCurrentThreadPriority(true, 99);
+    // Start AdvantageKit logger
+    Logger.start();
 
-        CommandScheduler.getInstance().run();
+    robotContainer = new RobotContainer();
 
-        Threads.setCurrentThreadPriority(false, 10);
+    FollowPathCommand.warmupCommand().schedule();
+    PathfindingCommand.warmupCommand().schedule();
+  }
+
+  /** This function is called periodically during all modes. */
+  @Override
+  public void robotPeriodic() {
+    Threads.setCurrentThreadPriority(true, 99);
+
+    CommandScheduler.getInstance().run();
+
+    Threads.setCurrentThreadPriority(false, 10);
+  }
+
+  /** This function is called once when the robot is disabled. */
+  @Override
+  public void disabledInit() {}
+
+  /** This function is called periodically when disabled. */
+  @Override
+  public void disabledPeriodic() {
+    robotContainer.updateDashboardStatus();
+  }
+
+  /** This autonomous runs the autonomous command selected by your {@link RobotContainer} class. */
+  @Override
+  public void autonomousInit() {
+    if (!matchStartingMethodCalled) {
+      matchStartingMethodCalled = true;
+      robotContainer.containerMatchStarting();
     }
 
-    /** This function is called once when the robot is disabled. */
-    @Override
-    public void disabledInit() {
+    autoCommand = robotContainer.getAutoCommand();
+    if (autoCommand != null) {
+      autoCommand.schedule();
     }
 
-    /** This function is called periodically when disabled. */
-    @Override
-    public void disabledPeriodic() {
-        robotContainer.updateDashboardStatus();
+    robotContainer.autoInit();
+  }
+
+  /** This function is called periodically during autonomous. */
+  @Override
+  public void autonomousPeriodic() {}
+
+  /** This function is called once when teleop is enabled. */
+  @Override
+  public void teleopInit() {
+    if (!matchStartingMethodCalled) {
+      matchStartingMethodCalled = true;
+      robotContainer.containerMatchStarting();
+    }
+    if (autoCommand != null) {
+      autoCommand.cancel();
     }
 
-    /**
-     * This autonomous runs the autonomous command selected by your
-     * {@link RobotContainer} class.
-     */
-    @Override
-    public void autonomousInit() {
-        if (!matchStartingMethodCalled) {
-            matchStartingMethodCalled = true;
-            robotContainer.containerMatchStarting();
-        }
+    robotContainer.teleopInit();
+  }
 
-        autoCommand = robotContainer.getAutoCommand();
-        if (autoCommand != null) {
-            autoCommand.schedule();
-        }
+  /** This function is called periodically during operator control. */
+  @Override
+  public void teleopPeriodic() {}
 
-        robotContainer.autoInit();
-    }
+  /** This function is called once when test mode is enabled. */
+  @Override
+  public void testInit() {}
 
-    /** This function is called periodically during autonomous. */
-    @Override
-    public void autonomousPeriodic() {
-    }
+  /** This function is called periodically during test mode. */
+  @Override
+  public void testPeriodic() {}
 
-    /** This function is called once when teleop is enabled. */
-    @Override
-    public void teleopInit() {
-        if (!matchStartingMethodCalled) {
-            matchStartingMethodCalled = true;
-            robotContainer.containerMatchStarting();
-        }
-        if (autoCommand != null) {
-            autoCommand.cancel();
-        }
+  /** This function is called once when the robot is first started up. */
+  @Override
+  public void simulationInit() {}
 
-        robotContainer.teleopInit();
-    }
-
-    /** This function is called periodically during operator control. */
-    @Override
-    public void teleopPeriodic() {
-    }
-
-    /** This function is called once when test mode is enabled. */
-    @Override
-    public void testInit() {
-    }
-
-    /** This function is called periodically during test mode. */
-    @Override
-    public void testPeriodic() {
-    }
-
-    /** This function is called once when the robot is first started up. */
-    @Override
-    public void simulationInit() {
-    }
-
-    /** This function is called periodically whilst in simulation. */
-    @Override
-    public void simulationPeriodic() {
-        // Update the simulation state
-        robotContainer.updateSimulation();
-    }
+  /** This function is called periodically whilst in simulation. */
+  @Override
+  public void simulationPeriodic() {
+    // Update the simulation state
+    robotContainer.updateSimulation();
+  }
 }
