+++ src/main/java/frc/robot/subsystems/rollers/Rollers.java
@@ -6,6 +6,9 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.subsystems.rollers.intake.Intake;
 import org.littletonrobotics.junction.Logger;
 
+import frc.robot.subsystems.rollers.sensors.*;
+
+import frc.robot.lib.generic_subsystems.rollers.*;
 public class Rollers extends SubsystemBase {
 
   public enum RollerState {
