+++ src/main/java/frc/robot/subsystems/rollers/intake/IntakeIO.java
@@ -1,5 +1,5 @@
 package frc.robot.subsystems.rollers.intake;
 
-import frc.robot.subsystems.rollers.GenericRollersIO;
+import frc.robot.lib.generic_subsystems.rollers.*;
 
 public interface IntakeIO extends GenericRollersIO {}
