+++ src/main/java/frc/robot/subsystems/rollers/intake/IntakeIOTalonFX.java
@@ -2,7 +2,7 @@ package frc.robot.subsystems.rollers.intake;
 
 import static frc.robot.subsystems.rollers.intake.IntakeConstants.*;
 
-import frc.robot.subsystems.rollers.GenericRollersIOTalonFX;
+import frc.robot.lib.generic_subsystems.rollers.*;
 
 public class IntakeIOTalonFX extends GenericRollersIOTalonFX implements IntakeIO {
