+++ src/main/java/frc/robot/subsystems/rollers/sensors/RollerSensorsIOComp.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.rollers;
+package frc.robot.subsystems.rollers.sensors;
 
 import edu.wpi.first.wpilibj.DigitalInput;
